requirements.txt
tensorflow
<<<<<<< HEAD
pandas 
=======
pandas
>>>>>>> 03dd003 (first commit)
dvc
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .